% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aluejaot.R
\docType{data}
\name{aluejaot}
\alias{aluejaot}
\title{Aluejaot ja aluetyypit}
\format{
A data frame with 309 rows and 11 variables:
\describe{
  \item{id}{Three number municipality code}
  \item{Kunta}{Municipality name}
  \item{NUTS 1}{}
  \item{Tukialue}{}
  \item{ELY-keskus}{}
  \item{Suuralue}{}
  \item{Maakunta}{}
  \item{Seutukunta}{}
  \item{Leader}{}
  \item{Alueluokka}{}
  \item{Kuntaryhma}{}
}
}
\source{
Alueet ja aluetyypit, Tilastokeskus,
 <http://stat.fi/media/uploads/tup/msind/msindaluejaot10.xls>
}
\usage{
aluejaot
}
\description{
Finnish municipalities with all the upper level regions
}
\examples{
aluejaot


}
\keyword{datasets}
